requirements.txt
altgraph==0.17.4
customtkinter==5.2.2
darkdetect==0.8.0
mysql-connector==2.2.9
mysql-connector-python==8.4.0
packaging==24.0
pillow==10.3.0
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
